Added explicit workflow and approval steps
In configuring the CircleCI config document, made a change that will add an approval step within CircleCI's UI/API prior to running last step.

Though generic at this point, the purpose appears to be to allow for test validations *prior* to allowing the commit to be merged into the branch.