Updated models to better define relationships
Completely revamped Sequelize model to more precisely articulate the relationships.

Goal is to unblock the I/O on sync; use non-default keys (i.e. listingId, not listing_id, where listingId is the name of the PK on the table listings, versus listing_id, where listing is the table name and id is the PK).