Removed undesired foreign keys
By eliminating the `hasMany` relationships, Sequelize no longer adds a [model]_[primary key] field to each model.

We have handled these relationships directly and explicitly in the model definition itself.

This approach *may* have shortcomings when it comes to method access and/or creating multiple records at once. It is not yet clear if this is the case.