Updated server to handle CORs; App to server dynamically by URL; new launch scripts
The app.use middleware seems to appropriately handle CORS wehreas the app.options with a headers object does not.

The URL to reach a page is now a get for listing/:listingId and then the React app will fetch the available nights using a separate route.

Also added new launch scripts for greater clarity in the package.json.